Refactor App component into smaller components
