+++ src/app/js/src/App.jsx
@@ -1,0 +1,182 @@
+import React, { Component } from 'react';
+import { arrayOf, bool, func, object, string } from 'prop-types';
+import { connect } from 'react-redux';
+
+import axios from 'axios';
+
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import Snackbar from 'material-ui/Snackbar';
+import { Row, Col } from 'react-flexbox-grid';
+
+import gistRequest from '../data/request.json';
+import {
+    clearScreen,
+    changeShareLink,
+    toggleShareSnackbarOpen,
+    toggleErrorSnackbarOpen,
+    toggleAddLayerDialog,
+} from './actions';
+import {
+    baseLayer,
+    map,
+    getDefaultTileJSON,
+    getBaseLayerTileJSON,
+} from './constants';
+import AddLayerDialog from './AddLayerDialog';
+import NavBar from './NavBar';
+import SideBar from './SideBar';
+
+class App extends Component {
+    constructor(props) {
+        super(props);
+        this.layers = [
+            baseLayer,
+        ];
+        this.clearLayers = this.clearLayers.bind(this);
+        this.share = this.share.bind(this);
+        this.handleShareSbRequestClose = this.handleShareSbRequestClose.bind(this);
+        this.handleErrorSbRequestClose = this.handleErrorSbRequestClose.bind(this);
+        this.openAddLayerDialog = this.openAddLayerDialog.bind(this);
+        this.addLayer = this.addLayer.bind(this);
+        this.removeLayers = this.removeLayers.bind(this);
+    }
+
+    componentDidMount() {
+        map.setTarget('map');
+    }
+
+    addLayer(layer) {
+        map.addLayer(layer);
+        this.layers.push(layer);
+        return this.layers.length;
+    }
+
+    removeLayers() {
+        for (let i = 0; i < this.layers.length; i += 1) {
+            if (this.layers[i] !== baseLayer) {
+                map.removeLayer(this.layers[i]);
+            }
+        }
+        this.layers = [
+            baseLayer,
+        ];
+    }
+
+    clearLayers() {
+        for (let i = 0; i < this.layers.length; i += 1) {
+            if (this.layers[i] !== baseLayer) {
+                map.removeLayer(this.layers[i]);
+            }
+        }
+        this.layers = [
+            baseLayer,
+        ];
+        this.props.dispatch(clearScreen());
+        const tileJSONList = getDefaultTileJSON();
+        if (document.getElementById('jsonTextarea')) {
+            document.getElementById('jsonTextarea').value = JSON.stringify(tileJSONList, null, '\t');
+        }
+    }
+
+    share() {
+        const tileJSON = this.props.tileJSON;
+        tileJSON.unshift(getBaseLayerTileJSON());
+        gistRequest.files['tile.json'].content = JSON.stringify(tileJSON, null, '\t');
+        axios.post('https://api.github.com/gists', gistRequest)
+            .then((response) => {
+                this.props.dispatch(changeShareLink({
+                    shareLink: `http://bl.ocks.org/d/${response.data.id}/`,
+                }));
+                this.props.dispatch(toggleShareSnackbarOpen({
+                    shareSnackbarOpen: true,
+                }));
+            });
+    }
+
+    handleShareSbRequestClose() {
+        this.props.dispatch(toggleShareSnackbarOpen({
+            shareSnackbarOpen: false,
+        }));
+    }
+
+    handleErrorSbRequestClose() {
+        this.props.dispatch(toggleErrorSnackbarOpen({
+            errorSnackbarOpen: false,
+        }));
+    }
+
+    openAddLayerDialog() {
+        this.props.dispatch(toggleAddLayerDialog({
+            showAddLayerDialog: true,
+        }));
+    }
+
+    render() {
+        const shareSnackbarMessage = (
+            <a className="snackbarLink" href={this.props.shareLink}>{this.props.shareLink}</a>
+        );
+        const errorSnackbarStyle = {
+            backgroundColor: '#fd4582',
+        };
+        let bar;
+        if (this.props.isCollapsed) {
+            bar = (
+                <NavBar
+                    openAddLayerDialog={this.openAddLayerDialog}
+                    share={this.share}
+                    clearLayers={this.clearLayers}
+                />
+            );
+        } else {
+            bar = (
+                <SideBar
+                    openAddLayerDialog={this.openAddLayerDialog}
+                    share={this.share}
+                    clearLayers={this.clearLayers}
+                    addLayer={this.addLayer}
+                    removeLayers={this.removeLayers}
+                />
+            );
+        }
+        return (
+            <MuiThemeProvider>
+                <div>
+                    <Row>
+                        {bar}
+                        <Col xs={this.props.isCollapsed ? 12 : 8} id="map" className="map" />
+                        <Snackbar
+                            open={this.props.shareSnackbarOpen}
+                            message={shareSnackbarMessage}
+                            onRequestClose={this.handleShareSbRequestClose}
+                        />
+                        <Snackbar
+                            open={this.props.errorSnackbarOpen}
+                            message={this.props.tileJSONParseError}
+                            onRequestClose={this.handleErrorSbRequestClose}
+                            bodyStyle={errorSnackbarStyle}
+                        />
+                        <AddLayerDialog
+                            addLayer={this.addLayer}
+                        />
+                    </Row>
+                </div>
+            </MuiThemeProvider>
+        );
+    }
+}
+
+App.propTypes = {
+    dispatch: func.isRequired,
+    tileJSON: arrayOf(object).isRequired,
+    shareLink: string.isRequired,
+    tileJSONParseError: string.isRequired,
+    shareSnackbarOpen: bool.isRequired,
+    errorSnackbarOpen: bool.isRequired,
+    isCollapsed: bool.isRequired,
+};
+
+function mapStateToProps(state) {
+    return state.main;
+}
+
+export default connect(mapStateToProps)(App);
